services/performance-controller/requirements.txt
# HTTP client for Prometheus and Simulation API
httpx==0.27.0

# Settings management
pydantic==2.10.4
pydantic-settings==2.7.0

# HTTP API for health checks and status
fastapi==0.115.6
uvicorn==0.34.0

# OpenTelemetry core
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-exporter-otlp==1.39.1
# Auto-instrumentation
opentelemetry-instrumentation-fastapi==0.60b1
